+++ argue/doc/changelog.rst
@@ -2,17 +2,19 @@
 Changelog
 =========
 
-v 0.1.0
+v0.1.0
+======
 
 Initial release! This is mostly a proof of concept initial implementation. It
 lacks several features I would consider required but works pretty well to start
 using it.
 
 Features:
+
 * argparse-like syntax
 * type-safe implementations for ``store``, ``store_const``, ``help``, and
   ``version`` actions
-* support for scalar (nargs=<default>, nargs='?') or
+* support for scalar ``(nargs=<default>, nargs='?')`` or
   container (nargs=<n>, nargs='+', nargs='*') assignment
 * provides different exception types and error messages for exceptional
   conditions so they can be distinguised between input errors (user errors),
@@ -20,3 +22,7 @@ Features:
 * support for custom actions
 * output formatting for usage, version, and full help text
 
+v0.1.1
+====================
+
+* Implemented subparser support